Reorganizing test files into their own classes along with their test files
