--- test/ecrypt_test.rb
@@ -1,11 +1,0 @@
-require 'minitest/autorun'
-require 'minitest/pride'
-require './lib/enigma'
-require './lib/date'
-require './lib/encrypt'
-require './lib/decrypt'
-require 'pry'
-
-class EnigmaTest < Minitest::Test
-
-end
